changed form handling and extracted form to a different file
