+++ src/App.js
@@ -8,8 +8,8 @@ import Login from './components/Login/Login';
 import BugsList from './components/Bugs/BugList/BugsList';
 import BugItemDetail from './components/Bugs/BugItemDetail/BugItemDetail';
 import Sidebar from './components/sidebar/Sidebar';
-import AddNewBug from './components/Bugs/AddNewBug/AddNewBug';
-import UpdateBug from './components/Bugs/UpdateBug/UpdateBug';
+import AddNewBug from './components/Bugs/AddNewBug';
+import UpdateBug from './components/Bugs/UpdateBug';
 import './store/firebase';
 
 function App() {
