--- src/components/Bugs/AddNewBug/AddNewBug.module.css
@@ -1,68 +1,0 @@
-/* CONTAINER */
-.container {
-  margin: 22px 5%;
-  display: flex;
-  flex-direction: column;
-  align-content: center;
-  justify-content: flex-start;
-}
-
-/* TITLE */
-.container h1 {
-  font-size: 40px;
-  font-weight: 900;
-}
-
-/* FORM */
-.bug-form {
-  width: 90%;
-  margin: 30px 5%;
-}
-
-/* EACH INPUT */
-.add-bug-input,
-.add-bug-textarea {
-  display: grid;
-  grid-template-columns: 100px 1fr;
-  padding: 5px 0;
-}
-
-.add-bug-input {
-  height: 50px;
-  line-height: 50px;
-}
-
-.add-bug-textarea {
-  height: 100px;
-  line-height: 100px;
-}
-
-.add-bug-input label,
-.add-bug-textarea label {
-  grid-column: 1;
-  font-size: 20px;
-  color: black;
-  font-weight: 500;
-  text-align: end;
-  padding-right: 10px;
-  border: none;
-}
-
-.add-bug-input input,
-.add-bug-input select,
-.add-bug-textarea textarea {
-  grid-column: 2;
-  border: none;
-  border-left: 2px solid var(--darkgray);
-  border-bottom: 1px solid var(--lightgray);
-  padding: 5px 0 5px 5px;
-  font-size: 18px;
-}
-
-/* ADD BUG BTN */
-.add-bug-btn {
-  background-color: var(--blue);
-  color: white;
-  position: relative;
-  top: 50px;
-}
