+++ src/components/Bugs/BugForm/BugForm.module.css
@@ -40,9 +40,11 @@
 .add-bug-input label,
 .add-bug-textarea label {
   grid-column: 1;
-  font-size: 18px;
+  font-size: 20px;
   color: black;
-  text-align: center;
+  font-weight: 500;
+  text-align: end;
+  padding-right: 10px;
   border: none;
 }
 
@@ -52,13 +54,13 @@
   grid-column: 2;
   border: none;
   border-left: 2px solid var(--darkgray);
-  border-bottom: 1px solid var(--darkgray);
+  border-bottom: 1px solid var(--lightgray);
   padding: 5px 0 5px 5px;
   font-size: 18px;
 }
 
-/* UPDATE BUG BTN */
-.update-bug-btn {
+/* SUBMIT BUG BTN */
+.submit-btn {
   background-color: var(--blue);
   color: white;
   position: relative;
